Partial Fix: Problem loading file saved with ADD:ABILITY and Craft skill
This partial fix prevents the dual popup of choosers problem, and it does properly make the association while the PC is in memory.  It also persists it to the PCG file (although it uses KEY: today and that KEY: doesn't identify the TYPE of object (e.g. TEMPLATE) to which that key applies, so the PCG is incomplete).

However, the system still requires abilities get cloned for virtual abilities, so that cloning still breaks the parsing since the association is not properly made during PCG load

Issue#: CODE-1117


git-svn-id: https://pcgen.svn.sourceforge.net/svnroot/pcgen/Trunk/pcgen@15796 b6a209e6-650d-0410-972b-dadffa731f64